back/license-supervisor/requirements.txt
bs4==0.0.1
requests==2.23.0
psycopg2-binary==2.8.5